+++ CHANGELOG.md
@@ -4,4 +4,5 @@
 
 ### Features
 - Initial commit with single socket instance with 3 NICs, creating full vpc
+- Requires 0.0.10 of catonetworks/vsocket-aws-ha-vpc/cato module